pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "jupyter (>=1.1.1,<2.0.0)",
    "dvc (>=3.59.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "matplotlib (<3.8)",
    "seaborn (>=0.13.2,<0.14.0)",
    "shap (>=0.47.2,<0.48.0)",
    "optuna (>=4.3.0,<5.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
flake8-isort = "^6.1.2"


[tool.flake8]
max-line-length = 80
ignore = [
    "WPS306",  # Разрешить использование классов с публичными атрибутами
    "WPS114",  # Разрешить короткие имена переменных в тестах
]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "src/venv",
    "*/venv/*",  # Игнорировать любые папки с именем venv
    "*/.venv/*",  # Игнорировать любые папки с именем .venv
    "dvc_storage",
    "notebooks",
    "*.egg-info",
    "*.eggs",
    ".mypy_cache",
    ".pytest_cache",
    ".tox",
    ".nox",
    "build",
    "dist",
]